Revert "getting ready for a test"
This reverts commit 1cc380eae0bf18c1252c93a1ba355c7b70832cef.

+++ src/main/java/glmath/glm/vec/_2/d/Vec2d.java
@@ -6,6 +6,7 @@
 package glmath.glm.vec._2.d;
 
 /**
+ *
  * @author elect
  */
 public class Vec2d extends FuncRelational {
@@ -16,7 +17,7 @@ public class Vec2d extends FuncRelational {
     }
 
     public Vec2d(int x, int y) {
-        this(x, (double) y);
+        this((double) x, (double) y);
     }
 
     public Vec2d(double x, double y) {
